+++ R/zzz.R
@@ -3,7 +3,6 @@
 # Suppress R CMD check notes for ggplot2 NSE
 utils::globalVariables(c("Weight", "Feature", "Directionality", "PC1", "PC2", "Group"))
 
-#' @importFrom phyloseq otu_table sample_data
 NULL